remove { isDefault: true } from tui, just print help
